+++ src/main/java/pl/asie/computronics/integration/railcraft/DriverReceiverBox.java
@@ -5,9 +5,13 @@ import li.cil.oc.api.machine.Callback;
 import li.cil.oc.api.machine.Context;
 import li.cil.oc.api.network.ManagedEnvironment;
 import li.cil.oc.api.prefab.DriverTileEntity;
+import mods.railcraft.api.signals.SignalAspect;
 import mods.railcraft.common.blocks.signals.TileBoxReceiver;
 import net.minecraft.world.World;
 import pl.asie.computronics.integration.ManagedEnvironmentOCTile;
+
+import java.util.LinkedHashMap;
+import java.util.Locale;
 
 /**
  * @author Vexatos
@@ -32,6 +36,15 @@ public class DriverReceiverBox extends DriverTileEntity {
 			} else {
 				return new Object[] { null, "signal receiver box is locked" };
 			}
+		}
+
+		@Callback(doc = "This is a list of every available Signal Aspect in Railcraft", getter = true)
+		public Object[] aspects(Context c, Arguments a) {
+			LinkedHashMap<String, Integer> aspectMap = new LinkedHashMap<String, Integer>();
+			for(SignalAspect aspect : SignalAspect.VALUES) {
+				aspectMap.put(aspect.name().toLowerCase(Locale.ENGLISH), aspect.ordinal());
+			}
+			return new Object[] { aspectMap };
 		}
 	}