+++ src/main/java/pl/asie/computronics/integration/railcraft/ReceiverBoxPeripheral.java
@@ -4,10 +4,14 @@ import dan200.computercraft.api.lua.ILuaContext;
 import dan200.computercraft.api.lua.LuaException;
 import dan200.computercraft.api.peripheral.IComputerAccess;
 import dan200.computercraft.api.peripheral.IPeripheral;
+import mods.railcraft.api.signals.SignalAspect;
 import mods.railcraft.common.blocks.signals.TileBoxReceiver;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.world.World;
 import pl.asie.computronics.integration.CCTilePeripheral;
+
+import java.util.LinkedHashMap;
+import java.util.Locale;
 
 /**
  * @author Vexatos
@@ -32,7 +36,7 @@ public class ReceiverBoxPeripheral extends CCTilePeripheral<TileBoxReceiver> {
 
 	@Override
 	public String[] getMethodNames() {
-		return new String[] { "getSignal" };
+		return new String[] { "getSignal", "aspects" };
 	}
 
 	@Override
@@ -40,15 +44,26 @@ public class ReceiverBoxPeripheral extends CCTilePeripheral<TileBoxReceiver> {
 		int method, Object[] arguments) throws LuaException,
 		InterruptedException {
 		if(method < getMethodNames().length) {
-			TileBoxReceiver box = this.tile;
-			if(!box.isSecure()) {
-				int signal = box.getTriggerAspect().ordinal();
-				if(signal == 5) {
-					signal = -1;
+			switch(method){
+				case 0:{
+					TileBoxReceiver box = this.tile;
+					if(!box.isSecure()) {
+						int signal = box.getTriggerAspect().ordinal();
+						if(signal == 5) {
+							signal = -1;
+						}
+						return new Object[] { signal };
+					} else {
+						return new Object[] { null, "signal receiver box is locked" };
+					}
 				}
-				return new Object[] { signal };
-			} else {
-				return new Object[] { null, "signal receiver box is locked" };
+				case 1:{
+					LinkedHashMap<String, Integer> aspectMap = new LinkedHashMap<String, Integer>();
+					for(SignalAspect aspect : SignalAspect.VALUES) {
+						aspectMap.put(aspect.name().toLowerCase(Locale.ENGLISH), aspect.ordinal());
+					}
+					return new Object[] { aspectMap };
+				}
 			}
 		}
 		return null;