+++ changelog.md
@@ -1,8 +1,15 @@
 ==== 1.1.1 ====
 
+  * ADDED: More Railcraft Integration
+            - You can now get/set the track mode in Locomotive Tracks using ComputerCraft/OpenComputers!
+                - You can access `modes` in OpenComputers or `modes()` in ComputerCraft to get a table containing all possible modes
+            - You can now get/set the speed limit in Limiter Tracks using ComputerCraft/OpenComputers!
+            - Receiver Boxes now have an `aspects` table containing every Signal Aspect in Railcraft
+                - You can access it from ComputerCraft using `aspects()`
   * ADDED: More Waila Integration
             - Tape Drives now show their current state in Waila
-            - Locomotive Relays now show whether they are bound to a locomotive
+  * CHANGED: Moved to OpenComputers 1.4
+  * CHANGED: MAJOR code refactor, file size should be a LOT smaller.
   * FIXED: A pretty serious bug with the Cipher Block
   * FIXED: Waila integration
   * FIXED: Tape Drive's getState not properly returning the current state.